pyproject.toml
[project]
name = "glmocr-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery[redis]>=5.3.0",
    "fastapi>=0.111.0",
    "glmocr[layout]>=0.1.3",
    "gunicorn>=21.2.0",
    "numpy>=2.4.3",
    "pillow>=10.0.0",
    "pydantic-settings>=2.2.0",
    "pypdfium2>=4.30.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.9",
    "redis>=5.0.0",
    "uvicorn[standard]>=0.29.0",
]
